+++ golang/lib/networks/network.go
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2020 - present Kurtosis Technologies LLC.
- * All Rights Reserved.
- */
-
-package networks
-
-// Docs available at https://docs.kurtosistech.com/kurtosis-libs/lib-documentation
-type Network interface{}