+++ golang/testsuite/services_impl/api/api_service.go
@@ -8,7 +8,7 @@ package api
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/services"
+	"github.com/kurtosis-tech/kurtosis-client/golang/services"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"io/ioutil"