+++ golang/testsuite/services_impl/nginx_static/nginx_static_container_config_factory.go
@@ -1,7 +1,7 @@
 package nginx_static
 
 import (
-	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/services"
+	"github.com/kurtosis-tech/kurtosis-client/golang/services"
 	"strconv"
 )