+++ golang/lib/testsuite/test.go
@@ -5,9 +5,7 @@
 
 package testsuite
 
-import (
-	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/networks"
-)
+import "github.com/kurtosis-tech/kurtosis-client/golang/networks"
 
 // Docs available at https://docs.kurtosistech.com/kurtosis-libs/lib-documentation
 type Test interface {