+++ golang/testsuite/services_impl/datastore/datastore_service.go
@@ -7,7 +7,7 @@ package datastore
 
 import (
 	"fmt"
-	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/services"
+	"github.com/kurtosis-tech/kurtosis-client/golang/services"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"io/ioutil"