+++ golang/testsuite/testsuite_impl/network_partition_test/network_partition_test_.go
@@ -6,9 +6,9 @@
 package network_partition_test
 
 import (
-	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/core_api_bindings"
-	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/networks"
-	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/services"
+	"github.com/kurtosis-tech/kurtosis-client/golang/core_api_bindings"
+	"github.com/kurtosis-tech/kurtosis-client/golang/networks"
+	"github.com/kurtosis-tech/kurtosis-client/golang/services"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/testsuite"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/services_impl/api"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/services_impl/datastore"